Minor changes
In users component change order of buttons and align center the text. In board-details add an empty field when in the column has one task because when user drag that task away from this column he cannot drag back and put in column where the task was. Fix changing column for task using select tag.