Change return type of pipe to fix bug
Bug occur in filter menu component. When the user unmark a tag in one table, the tag is also unmark in the other table, but the tag is taken into account in filtering and the task with this tag appears in the table.
That's why pipe getTagFromPipe returns an object with a check field so that all tags are checked every time the table is changed