Fix bug: delete a column
Every column is in a column wrapper to be able to switch order of columns but in context-menu was passed incorrect id. The user wasn't able to delete a column wrapper with column or edit column name. To solve this problem I added a input property to context-menu called deleteId and this property is undefined as default. Then is passed the user can delete a column wrapper or edit because the context-menu component has a column id and column wrapper id