Remove column wrapper table from database
Remove unnecessary table and replace it by an order field in column table. It will simplify a relations between tables and make easier to change column order because now order field is update in two columns. Before this change, there was a many unnecessary actions, removing and adding column wrappers to the board